final version , includes more BCM variables, exports DB records
git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/balewski.correlator@2308 b9700a98-9937-0410-b777-f318d6eb5bff

+++ treeBasedCorrelator/linRegBlue_v4/oneRunReg.py
@@ -8,16 +8,14 @@
 import os
 from optparse import OptionParser
 import datetime
-print "START JAN"
+from JanQwOnlLib import *
 
+print "START ONE run segment"
 wrkDir="./out/"
 safeDir="./web/"
 destWeb="/group/qweak/www/html/onlineRegression/autoA/"
-#destWeb="/group/qweak/www/html/onlineRegression/B/"
-# for run<=7717 use  1011_archive
-hclogWeb="https://hallcweb.jlab.org/hclog/1012_archive/"
-motherCsv="mother_online_summary.csv"
-
+motherCsvMD="mother_online_MD.csv"
+motherCsvLumi="mother_online_LUMI.csv"
 
 # -- command line options
 usage = "usage: %prog [options]"
@@ -41,22 +39,10 @@ parser.add_option("-n","--tot_events",
                   help="number of events to process ", 
                   default = 10000)
 
-parser.add_option("-v","--verbose", 
-                  dest="VERBOSE", 
-                  action = "store_true", 
-                  help="print all variables, matrices, and results", 
-                  default = True)
-
 parser.add_option("-c","--cuts", 
                   dest="CUTS", 
                   action = "store_true", 
                   help="use custom cuts",
-                  default = False)
-
-parser.add_option("-m","--mock_regress", 
-                  dest="MOCK_REGRESS", 
-                  action = "store_true", 
-                  help="skip regression and rename back existing output directory, post on the web", 
                   default = False)
 
 #... END
@@ -66,21 +52,16 @@ parser.add_option("-m","--mock_regress",
 # Performance tuning parameters #
 #################################
 
-VERBOSE=options.VERBOSE
 RUN_NO=options.RUN_NO
 RUN_SEG=options.RUN_SEG
 TOT_EVE=options.TOT_EVE
-MOCK_REGRESS = options.MOCK_REGRESS
 CUTS=options.CUTS
 RUNiSEG="%d." %RUN_NO + "%03d" % RUN_SEG
 
-
-if VERBOSE:
-    print ">>> Executing with parameters:"
-    print ">>>  MOCK_REGRESS = %s " % MOCK_REGRESS
-    print ">>>  RUN.SEG = %s" % RUNiSEG 
-    print ">>>  TOT_EVE = %d" % TOT_EVE
-    print ">>>  custom CUTS = %s" % CUTS
+print ">>> Executing with parameters:"
+print ">>>  RUN.SEG = %s" % RUNiSEG 
+print ">>>  TOT_EVE = %d" % TOT_EVE
+print ">>>  custom CUTS = %s" % CUTS
 
 ################################
 #  exporting HTML line
@@ -88,6 +69,53 @@ if VERBOSE:
 
 def  exportHtml(inpFile,outFile,pdfURL) :
     print inpFile,"--> export HTML -->", outFile
+
+    # first gather information from web pages
+    
+    # .... get HClog URL
+    err,text,id,mmddyy,hhmm,urlHClog=HClog_r(str(RUN_NO))
+    unixTime=Db2UnixTime_r_s(str(RUN_NO),str(RUN_SEG))
+    target=whatTarget(urlHClog)
+
+    print "target(s)=",target
+    ioPath=wrkDir
+    miscVal=', '.join(target)+", "+unixTime+", "+mmddyy+" "+hhmm+", "+urlHClog+", "+text.replace(',',';')
+    print "ioPath=",ioPath
+
+    #..... generate CVS record for MD
+    cmd_string12 = "root -b -q prCsvRecordTwo.C'(0,"+RUNiSEG+',"%s","'%miscVal+ioPath+'","'+ioPath+"\")'"
+    print "exec12:%s" % cmd_string12
+    fin,fout=os.popen4(cmd_string12)
+    fullLog=fout.readlines()
+    print fullLog
+
+    print "..... generate DB record" 
+    cmd_string12a = "root -b -q prCsvRecordTwo.C'(2,"+RUNiSEG+',"%s","'%miscVal+ioPath+'","'+ioPath+"\")'"
+    print "exec12a:%s" % cmd_string12a
+    fin,fout=os.popen4(cmd_string12a)
+    fullLog=fout.readlines()
+    print fullLog
+
+
+    #..... generate CVS record for LUMIs
+    cmd_string12b = "root -b -q prCsvRecordTwo.C'(1,"+RUNiSEG+',"%s","'%miscVal+ioPath+'","'+ioPath+"\")'"
+    print "exec12b:%s" % cmd_string12b
+    fin,fout=os.popen4(cmd_string12b)
+    fullLog=fout.readlines()
+    print fullLog
+
+    fullLogL=' '.join(fullLog).split("#regresRMS")
+    regRmsStr=fullLogL[1]
+    print '1$$$$$$$$$$$$',regRmsStr
+
+    fullLogL=fullLogL[2].split("#bcmRMS")
+    bcmRmsStr=fullLogL[1]
+    print '2$$$$$$$$$$$$',bcmRmsStr
+
+    nonrRmsStr=fullLogL[2].split("#nonregresRMS")[1]
+    print '3$$$$$$$$$$$$',nonrRmsStr
+
+     
     f=open(outFile,'w')
     #print f
     f.write("<!--   new run  R%s ------------>\n" %RUNiSEG)
@@ -100,41 +128,19 @@ def  exportHtml(inpFile,outFile,pdfURL) :
     cmd= "grep \#seenEve "+inpFile+" | cut -f2 -d\ "
     fin,fout=os.popen4(cmd)
     seenEve=int(fout.read())
-    f.write("<td  rowspan=2 > %d " %totEve +"<br> %d\n" %seenEve)
-    # .... get regRMS
-    cmd= "grep \#regresRMS "+inpFile+" | cut -f2 -d,"
-    fin,fout=os.popen4(cmd)
-    regRmsStr=fout.read()
-    f.write("<th> regres "+ regRmsStr+"\n")
+    f.write("<td  rowspan=2 > %d "%seenEve  +"<br><font size=\"-1\">  %d\n" %totEve)
+    f.write("<td  rowspan=2 > "+ target[0]+"<br> <font size=\"-1\"> "+target[1]+","+target[2]+"\n")
+    f.write(regRmsStr+"\n")
+    f.write(bcmRmsStr+"\n")
     #.....  get Sorting date'
     now=datetime.datetime.now()
     #str(now)
     #print now
     f.write("<td  rowspan=2 >"+now.strftime("%Y-%m-%d %H:%M:%S\n"))
-
-    # .... get HClog URL  
-    cmd= "wget --no-check-certificate "+hclogWeb+"logdir_auto.html -O - | grep \"Run %d"%RUN_NO+"\" | tail -n  1 >" +wrkDir+"logHC"
-    print "cmd_wget=",cmd
-    os.system(cmd)
-    f1=open(wrkDir+"logHC","r")
-      
-    myURL1=f1.readlines()[-1]
-    f1.close()
-    #print "aa1 ",myURL1
-    # in rare case there are multiple URLs in the same line
-    myURL2=myURL1.split("<LI>")[-1]
-    print "aa2 ",myURL2
-    
-    myURL=myURL2.replace('HREF="','HREF="'+hclogWeb)
-    #print "bb",myURL
-    f.write("<td  rowspan=2 > "+ myURL+"\n")     
-
-    # .... get nonregr RMS
-    cmd= "grep \#inputRMS "+inpFile+" | cut -f2 -d,"
-    fin,fout=os.popen4(cmd)
-    nonrRmsStr=fout.read()
-    f.write("<tr align=right> <th> nonreg "+ nonrRmsStr+"\n")
-
+    f.write("<td  rowspan=2 >"+mmddyy+" "+hhmm+"\n")
+    f.write('<td  rowspan=2 ><font size="-1"> <a href="'+urlHClog+'"> '+  text+'</a>\n')     
+    f.write(nonrRmsStr+"\n")
+#mmddyy,hhmm,url
     f.close()
 
 ################################
@@ -188,28 +194,30 @@ def  abortMe(code):
     #.....  get Sorting date'
     now=datetime.datetime.now()
     
-    cmd_stringAb = 'echo "<tr> <td>R%s'%RUNiSEG+' <td><td colspan=6 > aborted linear  regression <td  colspan=4 > sorted: '+now.strftime("%Y-%m-%d %H:%M:%S")+'">'+safeDir+"R%s.html"%RUNiSEG
+    cmd_stringAb = 'echo "<tr> <td>R%s'%RUNiSEG+' <td><td colspan=5 > aborted linear  regression <td  colspan=8 > sorted: '+now.strftime("%Y-%m-%d %H:%M:%S")+'">'+safeDir+"R%s.html"%RUNiSEG
     print "execAb:%s" % cmd_stringAb
     os.system(cmd_stringAb)
 
     os._exit(1)
 
 #############  M A I N  ###############
-if VERBOSE:
-    print ">>> Executing code ..."
-#i=int(xxx)
+print ">>> Executing code ..."
 
-
-#os._exit(1)
-os.system("rm -rf outOld")
-os.system("mv out outOld")
+# define all final pathes , names ......
 outDirName="R%s/"%RUNiSEG
+pdfName="regR%s.pdf"%RUNiSEG
+pdfURL="<A HREF=\""+outDirName+pdfName+"\"> R%s"%RUNiSEG+"</A>" + "<br> <A HREF=\""+outDirName+"logS2\"> regr.log </A>"
+finHtml=safeDir+"R%s"%RUNiSEG+".html"
+
+#print pdfURL
+#os._exit(1)
 retCode=0
 
-if MOCK_REGRESS:
-    os.system("cp -rp web/%s out" % outDirName )
-else:
+if 1:
+    os.system("rm -rf outOld")
+    os.system("mv out outOld")
     retCode=doubleRegress()
+#os._exit(1)
 
 if (retCode ):
     abortMe(retCode)
@@ -218,8 +226,8 @@ if (retCode ):
 cmd_string4 = "root -b -q plCor.C'(0,\"R%s\")'" % RUNiSEG 
 print "exec4:%s" % cmd_string4
 os.system(cmd_string4+">& out/log4")
+#os._exit(1)
 
-pdfName="regR%s.pdf"%RUNiSEG
 
 cmd_string5 = "cat out/R%s_page*ps "% RUNiSEG +" | ps2pdf - out/"+pdfName
 print "exec5:%s" % cmd_string5
@@ -230,15 +238,14 @@ cmd_string6 = "rm -f out/R%s_page*ps "% RUNiSEG
 print "exec6:%s" % cmd_string6
 os.system(cmd_string6+">& out/log6")
  
-pdfURL="<A HREF=\""+outDirName+pdfName+"\"> R%s"%RUNiSEG+"</A>" + "<br> <A HREF=\""+outDirName+"logS2\"> log </A>"
-print pdfURL
 
-exportHtml("out/logS2", safeDir+"R%s"%RUNiSEG+".html",pdfURL)
+exportHtml("out/logS2",finHtml ,pdfURL)
 
 cmd_string7 = "mv out "+safeDir+outDirName 
 print "exec7:%s" % cmd_string7
 os.system(cmd_string7)
 
+#..... order all HTML in reverse order
 cmd_string8 = "ls -r "+safeDir+"R*html "
 print "exec8:%s" % cmd_string8
 fin,fout=os.popen4(cmd_string8)
@@ -247,44 +254,30 @@ line=fout.read()
 line2=line.replace("\n"," ");
 #print "cc1 ",line2
 
-cmd_string9 = "rm -f "+safeDir+"index.html; cat "+safeDir+"top.html "+line2+safeDir+"bottom.html > "+safeDir+"index.html "
+cmd_string9 = "rm -f "+safeDir+"index.html; cat "+safeDir+"top.html "+line2+" "+safeDir+"bottom.html > "+safeDir+"index.html "
 #print "exec9:%s" % cmd_string9
 os.system(cmd_string9)
+#os._exit(1)
 
-#..... generate mothr CVS table
-print "rr ", RUNiSEG
-run=RUNiSEG[:4]
-seg=int(RUNiSEG[5:])
-print run, seg
-cmd="mysql --host=cdaql6.jlab.org --user=tracker data_tracker_qweak -e \"select   filename,start_time from total  where run_number=%s"%run+" and segment=%d"%seg+"\" |grep dat"
-print "exec DB time query:%s" % cmd
-fin,fout=os.popen4(cmd)
-string=fout.read()
-#print string
-unixTime=string.split()[1]
-print "R"+run+".%03d"%seg+ " "+ unixTime
 
-ioPath=safeDir+"R%s"%RUNiSEG+"/"
-print "ioPath=",ioPath
-cmd_string12 = "root -b -q prCsvRecord.C'("+RUNiSEG+',%s ,"'%unixTime+ioPath+'","'+ioPath+"\")'"
-print "exec12:%s" % cmd_string12
-os.system(cmd_string12)
-
-#...  merrge all CSV in to one big table
-cmd_string13 = "rm -f "+safeDir+motherCsv+"; cat "+safeDir+"R*/*csv  > "+safeDir+motherCsv
+#...  merrge all MD CSV in to one big table
+cmd_string13 = "rm -f "+safeDir+motherCsvMD+"; cat "+safeDir+"R*/*MD.csv  > "+safeDir+motherCsvMD
 print "exec13:%s" % cmd_string13
 os.system(cmd_string13)
 
-
+#...  merrge all LUMI CSV in to one big table
+cmd_string14 = "rm -f "+safeDir+motherCsvLumi+"; cat "+safeDir+"R*/*LUMI.csv  > "+safeDir+motherCsvLumi
+print "exec14:%s" % cmd_string14
+os.system(cmd_string14)
 
 
 #...... copy all to final web page
-cmd_string100 = "scp -rp -i ~/balewski/keys/id_rsa-ifarml4  "+safeDir+"/index.html  "+safeDir+outDirName+" "+safeDir+motherCsv+" balewski@ifarml4:"+destWeb
+cmd_string100 = "scp -rp -i ~/balewski/keys/id_rsa-ifarml4  "+safeDir+"/index.html  "+safeDir+outDirName+" "+safeDir+motherCsvMD+" "+safeDir+motherCsvLumi+" balewski@ifarml4:"+destWeb
 
 print "exec100:%s" % cmd_string100
 os.system(cmd_string100)
 os.system("date")
-
+os._exit(1)#tmp
 
 #..... regenarate plots, after CSV table is rebuild
 cmd_string110 = "(./doAutoPlots.sh)"
